backend_app/requirements.txt
flask
flask_restful
azure-data-tables
python-dotenv
pandas